# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Troostwatch API",
    "version": "0.6.2"
  },
  "paths": {
    "/lots": {
      "get": {
        "summary": "List Lots",
        "operationId": "list_lots_lots_get",
        "parameters": [
          {
            "name": "auction_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Auction Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LotView"
                  },
                  "title": "Response List Lots Lots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/positions/batch": {
      "post": {
        "summary": "Upsert Positions",
        "operationId": "upsert_positions_positions_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buyers": {
      "get": {
        "summary": "List Buyers",
        "operationId": "list_buyers_buyers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BuyerResponse"
                  },
                  "type": "array",
                  "title": "Response List Buyers Buyers Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Buyer",
        "operationId": "create_buyer_buyers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buyers/{label}": {
      "delete": {
        "summary": "Delete Buyer",
        "operationId": "delete_buyer_buyers__label__delete",
        "parameters": [
          {
            "name": "label",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Label"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sync": {
      "post": {
        "summary": "Trigger Sync",
        "operationId": "trigger_sync_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/live-sync/start": {
      "post": {
        "summary": "Start Live Sync",
        "operationId": "start_live_sync_live_sync_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiveSyncStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveSyncControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/live-sync/pause": {
      "post": {
        "summary": "Pause Live Sync",
        "operationId": "pause_live_sync_live_sync_pause_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveSyncControlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/live-sync/stop": {
      "post": {
        "summary": "Stop Live Sync",
        "operationId": "stop_live_sync_live_sync_stop_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveSyncControlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/live-sync/status": {
      "get": {
        "summary": "Get Live Sync Status",
        "operationId": "get_live_sync_status_live_sync_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveSyncStatusResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BuyerCreateRequest": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "BuyerCreateRequest"
      },
      "BuyerCreateResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "status",
          "label"
        ],
        "title": "BuyerCreateResponse"
      },
      "BuyerResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "BuyerResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LiveSyncControlResponse": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "LiveSyncControlResponse",
        "description": "Response for live sync control actions."
      },
      "LiveSyncStartRequest": {
        "properties": {
          "auction_code": {
            "type": "string",
            "title": "Auction Code"
          },
          "auction_url": {
            "type": "string",
            "title": "Auction Url"
          },
          "max_pages": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pages"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          },
          "interval_seconds": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Seconds",
            "description": "Seconds between sync runs; defaults to configured worker interval."
          }
        },
        "type": "object",
        "required": [
          "auction_code",
          "auction_url"
        ],
        "title": "LiveSyncStartRequest"
      },
      "LiveSyncStatusResponse": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "last_sync": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync"
          },
          "next_sync": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Sync"
          },
          "current_auction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Auction"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "LiveSyncStatusResponse",
        "description": "Status of the live sync worker."
      },
      "LotView": {
        "properties": {
          "auction_code": {
            "type": "string",
            "title": "Auction Code"
          },
          "lot_code": {
            "type": "string",
            "title": "Lot Code"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "current_bid_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Bid Eur"
          },
          "bid_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Count"
          },
          "current_bidder_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Bidder Label"
          },
          "closing_time_current": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Time Current",
            "description": "Current closing timestamp, if set."
          },
          "closing_time_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Time Original",
            "description": "Original closing timestamp, if set."
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": false
          },
          "effective_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Price"
          }
        },
        "type": "object",
        "required": [
          "auction_code",
          "lot_code"
        ],
        "title": "LotView",
        "description": "DTO representing a lot summary suitable for API responses."
      },
      "PositionBatchRequest": {
        "properties": {
          "updates": {
            "items": {
              "$ref": "#/components/schemas/PositionUpdate"
            },
            "type": "array",
            "title": "Updates"
          }
        },
        "type": "object",
        "required": [
          "updates"
        ],
        "title": "PositionBatchRequest"
      },
      "PositionBatchResponse": {
        "properties": {
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "created": {
            "type": "integer",
            "title": "Created",
            "default": 0
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "updated"
        ],
        "title": "PositionBatchResponse",
        "description": "Response for batch position updates."
      },
      "PositionUpdate": {
        "properties": {
          "buyer_label": {
            "type": "string",
            "title": "Buyer Label"
          },
          "lot_code": {
            "type": "string",
            "title": "Lot Code"
          },
          "auction_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auction Code"
          },
          "max_budget_total_eur": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Budget Total Eur"
          },
          "preferred_bid_eur": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Bid Eur"
          },
          "watch": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watch"
          }
        },
        "type": "object",
        "required": [
          "buyer_label",
          "lot_code"
        ],
        "title": "PositionUpdate"
      },
      "SyncRequest": {
        "properties": {
          "auction_code": {
            "type": "string",
            "title": "Auction Code"
          },
          "auction_url": {
            "type": "string",
            "title": "Auction Url"
          },
          "max_pages": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pages"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "auction_code",
          "auction_url"
        ],
        "title": "SyncRequest"
      },
      "SyncRunResultResponse": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "pages_scanned": {
            "type": "integer",
            "title": "Pages Scanned",
            "default": 0
          },
          "lots_scanned": {
            "type": "integer",
            "title": "Lots Scanned",
            "default": 0
          },
          "lots_updated": {
            "type": "integer",
            "title": "Lots Updated",
            "default": 0
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count",
            "default": 0
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SyncRunResultResponse",
        "description": "Result of a single sync run."
      },
      "SyncSummaryResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "auction_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auction Code"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncRunResultResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SyncSummaryResponse",
        "description": "Summary response for a sync operation."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}